Add task 3

+++ sandbox/src/main/java/ru/stqa/sandbox/Point.java
@@ -1,8 +1,8 @@
 package ru.stqa.sandbox;
 
 public class Point {
-  public double x;
-  public double y;
+  public int x;
+  public int y;
 
   public Point(int x, int y) {
     this.x = x;
